danggun_detail_Feat: #21 - danggun 수정, 삭제 controller 구현

+++ src/main/java/com/ounwan/controller/DanggunController.java
@@ -1,17 +1,24 @@
 package com.ounwan.controller;
 
+import java.io.IOException;
 import java.util.List;
+
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.ounwan.dto.ClientsDTO;
 import com.ounwan.dto.DanggunDTO;
 import com.ounwan.dto.ProductImagesDTO;
 import com.ounwan.service.DanggunService;
@@ -62,9 +69,38 @@ public class DanggunController {
 		return mv;
 	}
 
-	@RequestMapping(value = "/delete", method = RequestMethod.DELETE, produces="text/plain;charset=utf-8")
-	public @ResponseBody String danggunDelete(@RequestBody DanggunDTO danggun) {
-		int result = danggunService.deleteDanggun(danggun);
-		return (result == 1) ? "삭제 성공" : "삭제 실패";
+	@PostMapping(value = "/update")
+	public @ResponseBody boolean danggunUpdate(
+			@RequestPart(value = "imageFiles", required = false) MultipartFile[] imageFiles,
+			@RequestParam(value = "imageFilesNumber", required = false) int[] imageFilesNumber,
+			@RequestParam(value = "oldImageURL", required = false) String[] oldImageURL,
+			@RequestPart(value = "newDetailImages", required = false) MultipartFile[] newDetailImages,
+			@RequestParam("danggunNumber") Integer danggunNumber, 
+			@RequestParam("clientId") String clientId,
+			@RequestParam("name") String name, 
+			@RequestParam("price") Integer price,
+			@RequestParam("detail") String detail, 
+			@RequestParam("tradeHistoryNumber") Integer tradeHistoryNumber,
+			@RequestParam("imagesLength") int imagesLength,
+			@RequestParam("newImagesLength") int newImagesLength,
+			HttpSession session)
+			throws IllegalStateException, IOException {
+		ClientsDTO userInfo = (ClientsDTO) session.getAttribute("userInfo");
+
+		return danggunService.updateDanggun(imageFiles, imageFilesNumber, oldImageURL, newDetailImages,
+				userInfo.getClientId(), danggunNumber, clientId, name, price, detail, tradeHistoryNumber, imagesLength, newImagesLength);
+	}
+
+	@RequestMapping(value = "/delete", method = RequestMethod.DELETE, produces = "text/plain;charset=utf-8")
+	public @ResponseBody String danggunDelete(@RequestBody DanggunDTO danggun, HttpSession session) {
+		ClientsDTO clientInfo = (ClientsDTO) session.getAttribute("userInfo");
+		String clientId = clientInfo.getClientId();
+		boolean result;
+		if (clientId.equals(danggun.getClientId())) {
+			result = danggunService.deleteDanggun(danggun);
+		} else { // 다를 때
+			result = false;
+		}
+		return (result == true) ? "삭제 성공" : "삭제 실패";
 	}
 }
